Fix warnings; move header definitions.
